Moved exclusion classes to new namespace
