+++ FileWatcher/Configuration/Exclusions/Folders.cs
@@ -4,10 +4,10 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace TE.FileWatcher.Configuration
+namespace TE.FileWatcher.Configuration.Exclusions
 {
     /// <summary>
-    /// The files node in the XML file.
+    /// A folders node in the XML file.
     /// </summary>
-    public class Files : FileBase { }
+    public class Folders : FileBase { }
 }
